feat: Add support for v.redd.it
Also on Downloader.cs added sort of profile for each of the url. There
isn't a single group of params to fit all of the sources, some can
only output mp4 others only webm, and because the main objective
youtube-dl isn't be a converter it can't produce a single file type
for every profile for every source.